requirements.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python-headless
numpy
pytesseract
rembg
onnxruntime